refactor api and implementation of notifications
