+++ legend-depot-core-services/pom.xml
@@ -74,6 +74,10 @@
             <artifactId>mongo-java-server</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.finos.legend.depot</groupId>
+            <artifactId>legend-depot-store-notifications-api</artifactId>
+        </dependency>
 
     </dependencies>
