+++ legend-depot-store-notifications-impl/src/main/java/org/finos/legend/depot/store/notifications/NotificationsModule.java
@@ -21,11 +21,11 @@ import com.google.inject.Singleton;
 import org.finos.legend.depot.schedules.services.SchedulesFactory;
 import org.finos.legend.depot.store.notifications.api.Notifications;
 import org.finos.legend.depot.store.notifications.api.NotificationsManager;
+import org.finos.legend.depot.store.notifications.api.NotificationsStore;
 import org.finos.legend.depot.store.notifications.api.Queue;
 import org.finos.legend.depot.store.notifications.domain.QueueManagerConfiguration;
 import org.finos.legend.depot.store.notifications.resources.NotificationsManagerResource;
 import org.finos.legend.depot.store.notifications.services.NotificationsQueueManager;
-import org.finos.legend.depot.store.notifications.store.api.NotificationsStore;
 import org.finos.legend.depot.store.notifications.store.mongo.NotificationsMongo;
 import org.finos.legend.depot.store.notifications.store.mongo.NotificationsQueueMongo;
 import org.finos.legend.depot.store.notifications.store.mongo.NotificationsStoreImpl;
