+++ legend-depot-store-notifications-impl/src/main/java/org/finos/legend/depot/store/notifications/domain/QueueManagerConfiguration.java
