+++ legend-depot-store-notifications-impl/src/main/java/org/finos/legend/depot/store/notifications/resources/NotificationsManagerResource.java
@@ -18,14 +18,12 @@ package org.finos.legend.depot.store.notifications.resources;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import io.swagger.annotations.Example;
-import io.swagger.annotations.ExampleProperty;
 import org.finos.legend.depot.core.authorisation.api.AuthorisationProvider;
 import org.finos.legend.depot.core.authorisation.resources.BaseAuthorisedResource;
 import org.finos.legend.depot.domain.DatesHandler;
 import org.finos.legend.depot.domain.notifications.MetadataNotification;
 import org.finos.legend.depot.store.notifications.api.NotificationsManager;
-import org.finos.legend.depot.store.notifications.store.api.NotificationsStore;
+import org.finos.legend.depot.store.notifications.api.NotificationsStore;
 import org.finos.legend.depot.tracing.resources.ResourceLoggingAndTracing;
 
 import javax.inject.Inject;
