+++ legend-depot-store-server/pom.xml
@@ -53,7 +53,11 @@
         </dependency>
         <dependency>
             <groupId>org.finos.legend.depot</groupId>
-            <artifactId>legend-depot-store-notifications</artifactId>
+            <artifactId>legend-depot-store-notifications-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.finos.legend.depot</groupId>
+            <artifactId>legend-depot-store-notifications-impl</artifactId>
         </dependency>
         <dependency>
             <groupId>org.finos.legend.depot</groupId>
