+++ legend-depot-test-reports/pom.xml
@@ -101,7 +101,7 @@
         </dependency>
         <dependency>
             <groupId>org.finos.legend.depot</groupId>
-            <artifactId>legend-depot-store-notifications</artifactId>
+            <artifactId>legend-depot-store-notifications-impl</artifactId>
             <exclusions>
                 <exclusion>
                     <groupId>*</groupId>
